#5497 – Refactor (SnakeModePolymerBondRenderer): Use `SideChainConnectionBondRenderer.drawPartOfSideConnection(1)`
