+++ packages/ketcher-core/src/application/render/renderers/PolymerBondRenderer/SnakeModePolymerBondRenderer.ts
@@ -1,10 +1,7 @@
 import { editorEvents } from 'application/editor/editorEvents';
 import { CoreEditor } from 'application/editor/internal';
 import { Coordinates } from 'application/editor/shared/coordinates';
-import {
-  BaseMonomerRenderer,
-  BaseSequenceItemRenderer,
-} from 'application/render';
+import { SideChainConnectionBondRenderer } from 'application/render/renderers/PolymerBondRenderer/SideChainConnectionBondRenderer';
 import { D3SvgElementSelection } from 'application/render/types';
 import assert from 'assert';
 import { BaseMonomer, Vec2 } from 'domain/entities';
@@ -14,7 +11,6 @@ import {
   ConnectionDirectionInDegrees,
   ConnectionDirectionOfLastCell,
 } from 'domain/entities/canvas-matrix/Connection';
-import { CELL_WIDTH } from 'domain/entities/DrawingEntitiesManager';
 import { DrawingEntity } from 'domain/entities/DrawingEntity';
 import { PolymerBond } from 'domain/entities/PolymerBond';
 import { getSugarFromRnaBase } from 'domain/helpers/monomers';
@@ -220,72 +216,6 @@ export class SnakeModePolymerBondRenderer extends BaseRenderer {
   }
 
   // TODO: Specify the types.
-  private drawPartOfSideConnection({
-    cell,
-    connection,
-    direction,
-    horizontal,
-    sideConnectionBondTurnPoint,
-  }: {
-    readonly cell: Cell;
-    readonly connection: Connection;
-    readonly direction: ConnectionDirectionInDegrees;
-    readonly horizontal: boolean;
-    readonly sideConnectionBondTurnPoint: number;
-  }): {
-    readonly pathPart: string;
-    readonly sideConnectionBondTurnPoint: number;
-  } {
-    const sin = Math.sin((direction * Math.PI) / 180);
-    const cos = Math.cos((direction * Math.PI) / 180);
-    const xOffset = (CELL_WIDTH / 2) * cos;
-    const yOffset = (CELL_HEIGHT / 2) * sin;
-    const maxXOffset = cell.connections.reduce(
-      (max: number, connection: Connection): number => {
-        return max > connection.xOffset ? max : connection.xOffset;
-      },
-      0,
-    );
-    const maxYOffset = cell.connections.reduce(
-      (max: number, connection: Connection): number => {
-        const connectionYOffset = connection.yOffset || 0;
-        return max > connectionYOffset ? max : connectionYOffset;
-      },
-      0,
-    );
-
-    let endOfPathPart: number;
-    if (horizontal && sideConnectionBondTurnPoint) {
-      endOfPathPart = sideConnectionBondTurnPoint;
-    } else {
-      const { monomerSize, scaledMonomerPosition } = (
-        cell.monomer as BaseMonomer
-      ).renderer as BaseMonomerRenderer | BaseSequenceItemRenderer;
-      endOfPathPart = horizontal
-        ? scaledMonomerPosition.x + monomerSize.width / 2 + xOffset
-        : scaledMonomerPosition.y + monomerSize.height / 2 + yOffset;
-    }
-
-    const sideConnectionBondTurnPointInternal = endOfPathPart;
-
-    if (horizontal) {
-      endOfPathPart +=
-        -(connection.yOffset || 0) * 3 +
-        cos * -connection.xOffset * 3 +
-        cos * (maxXOffset + 1) * 3 +
-        (maxYOffset + 1) * 3;
-    }
-    let pathPart = horizontal ? 'H ' : 'V ';
-    pathPart += `${endOfPathPart - SMOOTH_CORNER_SIZE * cos} `;
-    pathPart += generateBend(cos, sin, cos, 1);
-
-    return {
-      pathPart,
-      sideConnectionBondTurnPoint: sideConnectionBondTurnPointInternal,
-    };
-  }
-
-  // TODO: Specify the types.
   private appendSideConnectionBond(rootElement, cells: Cell[]) {
     const firstCell = cells[0];
     const firstCellConnection = firstCell.connections.find(
@@ -364,7 +294,7 @@ export class SnakeModePolymerBondRenderer extends BaseRenderer {
         startPosition.x < this.sideConnectionBondTurnPoint
           ? 0
           : 180;
-      const result = this.drawPartOfSideConnection({
+      const result = SideChainConnectionBondRenderer.drawPartOfSideConnection({
         cell: firstCell,
         connection: firstCellConnection,
         direction,
@@ -448,13 +378,15 @@ export class SnakeModePolymerBondRenderer extends BaseRenderer {
           ? xDirection
           : // TODO?: Check. I am not sure about `as ConnectionDirectionInDegrees`.
             (previousConnection.direction as ConnectionDirectionInDegrees);
-        const result = this.drawPartOfSideConnection({
-          cell: previousCell,
-          connection: previousConnection,
-          direction,
-          horizontal,
-          sideConnectionBondTurnPoint: this.sideConnectionBondTurnPoint ?? 0,
-        });
+        const result = SideChainConnectionBondRenderer.drawPartOfSideConnection(
+          {
+            cell: previousCell,
+            connection: previousConnection,
+            direction,
+            horizontal,
+            sideConnectionBondTurnPoint: this.sideConnectionBondTurnPoint ?? 0,
+          },
+        );
         pathDAttributeValue += result.pathPart;
         this.sideConnectionBondTurnPoint = result.sideConnectionBondTurnPoint;
       }
